+++ sdk/sdk/src/riscv_ecalls/sha_extend.rs
@@ -5,11 +5,11 @@ use core::arch::asm;
 ///
 /// ### Safety
 ///
-/// The caller must ensure that `w` is valid pointer to data that is aligned along a four byte
-/// boundary.
+/// The caller must ensure that `w` is valid pointer to data that is aligned along an eight byte
+/// boundary. Each logical SHA word is taken from the low 32 bits of one `u64` slot.
 #[allow(unused_variables)]
 #[no_mangle]
-pub extern "C" fn syscall_sha256_extend(w: *mut [u32; 64]) {
+pub extern "C" fn syscall_sha256_extend(w: *mut [u64; 64]) {
     #[cfg(target_os = "zkvm")]
     unsafe {
         asm!(
